+++ src/TrashLib.TestLibrary/TestConfig.cs
@@ -1,0 +1,13 @@
+using JetBrains.Annotations;
+using TrashLib.Config.Services;
+
+namespace TrashLib.TestLibrary;
+
+[UsedImplicitly]
+public class TestConfig : ServiceConfiguration
+{
+    public TestConfig()
+    {
+        Name = "Test";
+    }
+}