+++ src/TrashLib/Config/Services/IServiceConfiguration.cs
@@ -2,6 +2,7 @@ namespace TrashLib.Config.Services;
 
 public interface IServiceConfiguration
 {
+    string Name { get; }
     string BaseUrl { get; }
     string ApiKey { get; }
     ICollection<CustomFormatConfig> CustomFormats { get; }